clamped residuals, switched torch.float16 to torch.bfloat16 to fix the overflow causing NAN, and also did a log transform to the mel spectograms to provide stability for AST since AST is pretrained on log-mel

+++ src/eval.py
@@ -79,7 +79,7 @@ def run_eval(model, dataloader, device: torch.device, max_batches: int = 0):
         tgt_output = tokens[:, 1:]
         padding_mask = tgt_input == PAD
 
-        with autocast(dtype=torch.float16):
+        with autocast(dtype=torch.bfloat16):
             outputs = model(mel, tgt_input, padding_mask)
             logits = outputs["logits"]
             loss = ce_loss_fn(
